Add FlowGraph.toJSON, refactor
