+++ com.ibm.wala.cast.js/src/main/java/com/ibm/wala/cast/js/callgraph/fieldbased/flowgraph/FlowGraph.java
@@ -59,6 +59,7 @@ import com.ibm.wala.util.collections.HashSetFactory;
 import com.ibm.wala.util.collections.Iterator2Iterable;
 import com.ibm.wala.util.collections.Pair;
 import com.ibm.wala.util.graph.Graph;
+import com.ibm.wala.util.graph.GraphPrint;
 import com.ibm.wala.util.graph.GraphReachability;
 import com.ibm.wala.util.graph.GraphSlicer;
 import com.ibm.wala.util.graph.NumberedGraph;
@@ -579,5 +580,9 @@ public class FlowGraph implements Iterable<Vertex> {
         return null;
       }
     };
+  }
+
+  public String toJSON() {
+    return GraphPrint.graphToJSON(graph);
   }
 }
